projects/pyproject.toml
[project]
name = "headhunter-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.12.0",
    "altair>=5.0.0",
    "black>=23.0.0",
    "faiss-cpu>=1.7.4",
    "jupyter>=1.0.0",
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-upstage>=0.1.0",
    "langgraph>=0.1.0",
    "matplotlib>=3.10.6",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "plotly>=5.17.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.0.0",
    "pypdf2>=3.0.1",
    "pytest>=7.4.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=2.2.0",
    "sqlalchemy>=2.0.0",
    "streamlit>=1.28.0",
    "streamlit-chat>=0.1.1",
    "streamlit-option-menu>=0.3.6",
    "tavily-python>=0.3.0",
    "torch>=2.0.0",
    "tqdm>=4.65.0",
    "transformers>=4.30.0",
]
